Move SBOM models to their own file
