+++ cachi2/core/package_managers/npm.py
@@ -15,7 +15,8 @@ from cachi2.core.checksum import ChecksumInfo, must_match_any_checksum
 from cachi2.core.config import get_config
 from cachi2.core.errors import PackageRejected, UnexpectedFormat, UnsupportedFeature
 from cachi2.core.models.input import Request
-from cachi2.core.models.output import Component, ProjectFile, Property, RequestOutput
+from cachi2.core.models.output import ProjectFile, RequestOutput
+from cachi2.core.models.sbom import Component, Property
 from cachi2.core.package_managers.general import async_download_files
 from cachi2.core.rooted_path import RootedPath
 from cachi2.core.scm import RepoID, clone_as_tarball, get_repo_id
